requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2024.12.0
google-auth==2.47.0
google-genai==1.60.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-mongodb==0.11.0
langchain-neo4j==0.8.0
langchain-openai==1.1.8
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-checkpoint-mongodb==0.3.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.6.9
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mem0ai==1.0.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==6.1.0
neo4j-graphrag==1.13.1
networkx==3.6.1
numpy==2.4.1
ollama==0.6.1
openai==2.15.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
portalocker==3.2.0
posthog==7.9.4
propcache==0.4.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.5
pymongo-search-utils==0.3.0
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
rank-bm25==0.2.2
redis==7.1.1
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
rich-toolkit==0.17.1
rignore==0.7.6
rq==2.6.1
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.1
sentry-sdk==2.50.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.5
SQLAlchemy==2.0.46
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.0.0
typer==0.21.1
typer-slim==0.21.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
